requirements.txt
django==5.0.6
djangorestframework==3.15.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
channels==4.1.0
django-cors-headers==4.3.1
daphne==3.0.2